backend/foodgram/requirements.txt
asgiref==3.7.2
Django==3.2.13
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
drf-extra-fields==3.7.0
filetype==1.2.0
Pillow==10.0.1
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
djoser==2.1.0
